Change search to GET request
