+++ views/index.jade
@@ -4,4 +4,4 @@ if req.loggedIn
   p
     a(href='/create-lesson', title='Makin lessons') Create Lesson
 p
-  a(href='/lessons', title='Search fa shit') Search
+  a(href='/search', title='Search fa shit') Search
